Add LVM module to OpenStack Common
*!!!! WIP !!!!*
*Needs Unit Tests*

This patch adds a simple LVM wrapper/interface to the
OpenStack-Common modules.  Currently we use LVM in a couple
of places, this would allow us to remove the duplication of
that code, particularly between Nova and Cinder.

More valuable however is the ability to utilize local storage
for compute instances as we go forward.

Implements Blueprint: add-lvm-module

Change-Id: I58dd70937110e50b5b5c49232e66f1c7ea82cbe6