Use the standard logging module
The oslo logging module brings in oslo.cfg
which is not desirable for projects that do
not want to bring it in. In these cases that
logging module is not adding any benefit over
the standard python logging module so just
use it directly instead.

This encourages small isolated code modules
that do not bring in large dependent modules.

Change-Id: I7c4c90fb43b63cef00bdb533872e6d257d761c10