Fix database connection string is secret
The database connection string was not marked as secret, so it
would be printed out in cleartext in the logs when config settings
were logged.

The database connection string typically contains the password
that's used to connect to the database, so it should be marked as
secret so that it doesn't get logged.

Change-Id: If6e1026814262c961a4f22d90f546cec34831bc8
Closes-Bug: #1266590