Add support for zh locales missing from babel
Babel>=1.0,<1.3 has a bug where 'zh_CN' and 'zh_TW' aren't supported
even though they are perfectly legitimate locales, see:
https://github.com/mitsuhiko/babel/issues/37

In Babel 1.3 they fixed the bug and they support these locales, but
they are sill not explicitly "listed" by locale_identifiers(). That's
why we add the locales if necessary so they are listed as supported
otherwise when validating supported locales for the transalated log, we
would fail to allow zh_TW and zh_CN even if there are .mo files for
them.

Change-Id: Ie61539a709808dfea690a59602be94347fcfbbb9