Add 'new' parameter mock.Patch and mock.PatchObject classes
These two classes are using mock module, although both of them weren't using the
'new' parameter from mock.patch and mock.patch.object which allow us to set a
specific method to be executed instead of the using MagicMock.

As mock's patch doc string says (and it says in patch,object that the parameters
are the same):
"If `new` is omitted, then the target is replaced with a 'MagicMock'."

Unit test to prove that both approaches work were also added.

Change-Id: I32a90b4d47174b6adb53b525a4add7002411b29d