Use entry points to discover options in libraries
When generating the sample configuration file, we want to include
options defined in libraries used by the application (especially
oslo libraries). Add a section to the sample generator to look
at entry points in the 'oslo.config.opts' namespace and include
their return values in the output.

Implements: blueprint improve-config-discovery-for-docs

Change-Id: I635cd8e60ca687851c9dd891667ecc2ca3e8720c