Allow different login methods to be used with kombu connections.
Currently it is not possible to modify the login method used with
kombu connections via configuration files, and requires modifying
the source code to actually add a login method in the first place.
This adds a configuration option (and default) which are passed to
the underlying kombu connection.

Unit tests have also been corrected to the best of my knowledge.

Closes-Bug: #1269867
Change-Id: I6ff421a65b0e6f93db76d91e9362cd63d9170299