RabbitMQ: don't start failover from the failed broker
In RabbitMQ implementation, when using multiple rabbit_hosts, we don't want to
immediately retry failed connection for the same failed broker. This was not
the case in existing implementation though, where we've always attempted to
reconnect starting from the first broker in the list of candidates. So if the
first broker failed, we initiated reconnect to the same failed broker.

This change makes reconnect() implementation to start from the broker which is
*after* the just-failed broker in the list of candidates. This is done for
failover mode only, so initial connection attempt will still start from the
first broker in the list.

Change-Id: I67923cb024bbd143edc8edccf35b9b400df31eb3
Partial-Bug: 1261631