Enhance tests for LOG.deprecated
The tests for LOG.deprecated were incorrect because they were
clearing the log buffers on every call.

When this is corrected, it shows that LOG.deprecated is not working
as expected since it logs multiple times when the test indicates
that it should only log once.

Also, added some new tests for logging with substitution arguments.

Change-Id: Ic0e4925c372e5948c2140daa07ab3362fb25a306
Partial-Bug: #1266812