Add missing entries to MAINTAINERS
This is required for the subsequent addition of a tool to verify
that all files in the oslo-incubator tree are covered by an entry
in MAINTAINERS.

For files that had only a single contributor, I made that person
the maintainer.  For modules with multiple contributors I set them
to orphaned.  Ideally those entries will be updated as appropriate
by the actual maintainers.

Change-Id: Ide60c101c37b17ce6324d1545c05698b287011e0