Add common code to matchers
Several projects use the same matchers for dicts and XML files,
leading to code duplication. So far, this is included in
nova, keystone, ironic and trove projects.

This is the nova matchers file pep8 clean. Other projects matchers,
are just a copy (with minimal divergences) or a subset of this one.

Change-Id: Id2d74a6eaf9aa295165b7f78586581e661beb91f
Partial-Bug: #1199536