Make it clear qpid_hosts order is not to be relied on
Qpid implementation may change the order of qpid_hosts selection. Make
it clear in rabbit_hosts option description.

Users should not expect any particular order of connection attempts.
F.e. users could probably expect that we will always start connection
attempts from the first host in the list, or that we'll switch them
for failover in round-robin order, or that on failover we'll always
restart from the first host in the list. All these options are
somewhat reasonable and are left for implementation to decide.

The change was suggested by Flavio Percoco in private IRC discussion.

Change-Id: If4aed820f954ea9e0323e76337f0e9cc164678e1