Introduce mysql_sql_mode option, remove misguided warning
Introduce a config option, mysql_sql_mode, that allows the user
to set the desired SQL mode for MySQL sessions.

For MySQL sessions, set the SQL mode session variable
accordingly.

For compatibility purposes, retain the mysql_traditional_mode
argument in method signatures, but only use it to override the
config option as if it were set to 'TRADITIONAL'.

SQL injection considerations: the value of mysql_sql_mode
is guarded via StrOpt "choices", an invalid or malicious value
would raise a ValueError for the config option. A malicious
application could still call _set_session_sql_mode with a value
that injects unwanted SQL.

Change-Id: Ia527e850671ab9542db5864617384cee21e91bf6
Closes-bug: 1271706