Introduce mysql_sql_mode option, remove old warning
Introduce a config option, mysql_sql_mode, that allows the user
to set the desired SQL mode for MySQL sessions.

For MySQL sessions, set the SQL mode session variable
accordingly.

For compatibility purposes, retain the mysql_traditional_mode
argument in method signatures, but only use it to override the
config option as if it were set to 'TRADITIONAL'.

Change-Id: Ia527e850671ab9542db5864617384cee21e91bf6
Closes-bug: 1271706