Don't store engine instances in oslo.db
oslo.db is meant to be a collection of helper utilities for
SQLAlchemy. In order to behave like a 'good' library, we must
stop using global state.

This patch ensures we don't store any engine instances globally in
oslo.db. It's up to end applications to decide how to cope with
engines, not oslo.db.

Partial-Bug: #1263908

Change-Id: I330467ec1317e1a13ff9f0237e2d8900d718e379