Don't use cfg.CONF in oslo.db
Using of a global config object is something we want to avoid when
making libraries. Parameters must be passed as function/class methods
arguments instead.

In order not to duplicate options declaration in each project, they
have been put into separate options module (which should not be a
part of oslo.db when it's released; we'll probably want to find a
place for it within oslo-incubator).

Partial-Bug: #1263908

Change-Id: Ib5a4e31b4e78e2b4cb807a8c88b8072f4207eb22