Make wrap_db_error() a class decorator
For a long time _wrap_db_error() had been a function decorator, which
used the global Engine instance to find out the DB backend name.

We don't have the global Engine instance now, so wrap_db_error() must
be able to access Session instance context instead. And it is meant
to be applied to Session methods only. Which means we can reasonably
assume, that the wrapper function always receives 'self' argument
pointing to a Session subclass instance. Though, it's obviously would
be better, if we could enforce that wrap_db_error() was applied only
to Session subclasses methods at the decorator application time
rather than at the runtime.

In Python 3 there is no way to distinguish between an 'unbound' class
method and a general function (as there are no unbound methods
anymore). So the cleaner solution must be making wrap_db_error()
a class decorator instead, which will raise TypeError when applied
to not a Session subclass.

Related-Bug: #1263908

Co-authored-by: Victor Sergeyev <vsergeyev@mirantis.com>

Change-Id: Ib3f5979c4e6e2ab98803eea9107cfe4c93bd2884