Make sessionmaker cope with a slave engine
oslo.db uses a notion of a 'slave' engine, which is an additional DB
connection. This connection can be used in order to decrease the load
on the master DB connection (all reads are passed to the slave DB
connection).

As we no longer store global engine instances in oslo.db, API must
be updated to make it possible to use a slave engine.

Related-Bug: #1263908

Change-Id: I39f92ab66c943e3762f6996f197ea42664f4b109