generator: use EXTRA_* env vars in the bash script
It's a bit odd to source a file (intended to contain env vars) from
the bash script and then for the python module be the one to actually
use the env vars, especially since the python module has command line
arguments.

Instead, lets read the values from the env vars in the bash script and
pass them to the python via command line arguments.

This is a slightly incompatible change - the env var was expected to
be comma separated before, but now its $IFS separated. However, none
of the users of this rc file actually have multiple values listed
in EXTRA_MODULES.

Change-Id: I79ce84e17c768daf13a7c8a3366e658ad10999b7