Enhance config fixture to support registering options
Add register_opt() and register_opts() to the config fixture. This
allows tests to define configuration options that will not persist
in the global CONF object over multiple tests.

Minor changes to the config fixture tests to use the same fixture
that the test case uses as .config_fixture instead of an
independantly (non setUp) instantiation of the fixture.

Change-Id: Ic125b705644ae091cd6696e7d8619a3238a8c866