Add tool for helping with library graduation
This new script prepares a copy of the oslo-incubator repository
to house a stand-alone oslo library by revising the history,
retaining commits for files that are part of the new library and
discarding commits for files that are not. It then uses the
new oslo-cookiecutter library to inject any missing structural
files so the packaging and tests will work.

Change-Id: I1f513a773589f034975ae6ee54c5a418e4bfce25