Return the unicode representation of an object
Objects in openstack are typically *safe* to translate to
unicode but for those that want to ensure that unicode is
returned for their object type (str() is not safe to call
to perform this in python2.x and python3.x). This utility
function will translate an object safely to unicode and
resort back to safe decoding for when this decoding is
problematic.

Change-Id: Ic996fdde13b28ebeb750289b81663ac3981d05ac