Move memorycache expiry out of .get method
The entire memory cache was being scanned for items which had passed
their timeout every time an item was fetched.  This could cause a
performance problem for large caches.

This patch still has the clean on the .get method, but will skip the
cache cleaning unless >60s has elapsed since the last time.  A
method is exposed for clients to call if they have a busy cache that
needs cleaning more often.

Closes-Bug: #1276530

Change-Id: Ibe92b20bb389a392d6a62093ca6305a9d492a64f