Isolate utcnow mocking in the ManagerTestCase
The mock of utcnow in the ManagerTestCase is affecting PeriodicTasksTestCase
that the test result of
"tox -epy26 unit.test_periodic" and
"tox -epy26 unit.test_periodic.PeriodicTasksTestCase" are different.

This patch isolates the mocking by removing set_time_override and
using the mock.patch decorator on each test cases which needs it.

Also this bug caused "doit_with_kwargs_odd" to be executed accidentally
eventhough the task spacing was 10 seconds. This was fixed
by changing the spacing to 1 and sleeping once between the two runs.

Change-Id: I3385c8d313d58cbaad5c21347b2b7ce4a8aae6a7
Closes-bug: #1277823