Reverts test condition in test_periodic
The run_immediately argument was added to "doit_with_kwargs_odd" in patch
47c9d606 as temporary workaround for the race condition fixed in patch
9b39d1d6.

However having this argument passed conceals a corner case which is caused by
the time gap between when the sciprts are loaded and when the tests
are actually conducted.

This phenomenon can be regenerated by changing spacing back to 10 on this
patch and running "tox -epy26" and "tox -epy26 unit.test_periodic" separately.

There is a test case for testing run_immediately in the ManagerTestCase so
this patch reverts parts of 47c9d606 of not passing a run_immediately argument
and increasing the spacing value.

Change-Id: I3385c8d313d58cbaad5c21347b2b7ce4a8aae6a7
Related-bug: #1275151
Related-bug: #1275152