Fixes race condition in test_periodic
The mock of utcnow in the ManagerTestCase is affecting PeriodicTasksTestCase
that the two test cases are in race condition.
This can be verified by logging timeutils.utcnow() with intentionally
creating an assertion error on line 72. Then "tox -epy26 unit.test_periodic"
and "tox -epy26 unit.test_periodic.PeriodicTasksTestCase" have
different results of utcnow().

This patch isolates the mocking by removing set_time_override and
using the mock.patch decorator on each test cases which needs it.

Change-Id: I3385c8d313d58cbaad5c21347b2b7ce4a8aae6a7
Closes-bug: #1277823