Support "hexdigest" output from HKDF.extract
Add in a keyword argument to HKDF.extract to specify (boolean)
if a hexdigest output is desired. If `hexdigest` evaluates to True,
`HMAC.new(...).hexdigest()` will be returned instead of
`HMAC.new(...).digest()`. This is useful in cases that the extracted
key is being used in a key-value-store or other system that only
supports a limited (e.g. ASCII) character set as the key.

Change-Id: I30c29ba1961047630ebcfb072430286484f74b45