Revert "Remove lazy loading of database backend"
This reverts commit 97d8cf4cbda843000f97ae1e3ca93de32003cf67.

We removed this a while ago, but unfortunately Nova, Ironic and
probably other projects still rely on this feature to prevent import
cycles and postpone initialization of DB backend to the moment when
oslo.config has already parsed the config files (the errors were
hidden when we tested the original commit by the fact that we had
lazy loading of engines in oslo.db code). Reverting is needed
in order to make transitioning to oslo.db as smooth as possible.

Change-Id: I9fe09c132c716345f1cef240483b023e850d4989