Add a base test case for DB schema comparison
Add a base test case that allows to check if the DB schema obtained
by applying of migration scripts is equal to the one produced from
models definitions. It's very important that those two stay in sync
to be able to use models definitions for generation of the initial
DB schema. Though, due to the using of sqlalchemy-migrate we have
many divergences in projects right now, which must be detected and
fixed.

Fortunately, we can now rely on alembic + sqlalchemy implementation
of schema comparison instead of writing our own tools.

Change-Id: I0758c7e09d1d46ce870c3f94f76c2a177955e143