Scrub passwords in the config section of GMRs
Currently, the config section of the Guru Meditation Reports
displays all the configuration options verbatim.  This means
that passwords will be dumped.  This can present a security risk
(as someone could potentially not have access to the config files,
but be allowed to dump/view the Guru Meditation Reports).

Ergo, a new parameter has been added to the Config generator and
model, scrub_passwords, which defaults to True.  When enabled,
any key of the form /^\w+_(password|token|key)$/ will have its
value masked with stars in the GMR.

Change-Id: Ide08e859465caf32bdc3a53ed07b24737a368d25