migration.db_sync requires an engine now
One of the migration test cases was still not passing an engine to
db_sync.  This doesn't show up in the gate because it is a negative
test that was failing with the correct exception for a different
reason.  I happened to find the problem because I had a file named
foo in my oslo-incubator directory from some previous testing and foo
was being passed in as the repository path, so that call wasn't
failing like it would on a clean system such as the gate.

Change-Id: Ia10cfbd4b99d821d9a725e6eb256fdd8ab001405