Look for well-defined options function
The config sample generator supports plugins for libraries, but that
does not work for some applications. In particular, keystone has a
well-organized dictionary of options that the discovery code does
not recognize. This change allows an app to provide a function with a
well-defined name which will be discovered and used to retrieve option
definitions, just like the plugins.

Change-Id: Ie4d5ea8c474d961aeaf8d12538a4ac394c231400