Fix import order in test_excutils
The tests.unit.test_excutils module would fail to import by
itself because it couldn't import the mox modules from six.moves.
This is because the mox module is registered with six.moves in
openstack.common's __init__ and that hadn't been imported yet.

The fix is to use mox from moxstubout since that part imported
mox already.

Change-Id: I786b9bffba28935f59da0dcf84ef2b5aa3f2b58d