Use http exceptions from Webob - aliases creation
All exceptions in apiclient.exceptions which related to HTTP are equal to
exceptions from webob.exc.
To decrease code duplication it is needed to remove our custom exceptions
and use http exceptions from WebOb library.

This is first step: Aliases creation for all custom exceptions which are
duplicated with webob.exc.

Change-Id: Ia1d3be8a81818252946c07e0ed5c9eca1725d999