Apply testscenarios to imageutils test scripts
There are some corner cases that need to be tested such as
processing floating values in imageutils.QemuImgInfo.
Rather than copy and pasting test scripts, it would be better
to have testscenarios applied for ease of addiing test cases.

The original test case was not removed to prove the new test script
covers the original tests.
The test_imageutils module was removed from py33 tests due to
testscenarios incompatibility.

Change-Id: I4aa71034a5f1e83724efb49ceec3054726e87233