Add from_config() method to EngineFacade
Add a helper that allows to create an EngineFacade instance given an
oslo.config instance. This helper is completely optional, but it was
requested by a few people syncing the latest oslo.db code to target
projects.

Note, that oslo.config is mocked in tests, so oslo.db still has no
runtime dependency on oslo.config.

Change-Id: Icab476386b859be0b0e984a94cbfcf3d581ab6d2