Reports: Use sig handler traceback for curr thread
Previously, when an application dumped a GMR, the traceback for
the thread in which the GMR code executed showed the GMR code
itself.  This is not particularly useful.

Thankfully, signal handlers are passed a Python frame object
which contains the traceback from just before the signal handler
is triggered.  We take this traceback and use it to replace
the traceback for the current thread in the threading generator.

While the GMR signal handler passes this on to the GMR class
and threading generator, the traceback parameter is optional
in both places.  When not supplied with a traceback, the GMR
and generator revert the the behavior from before this patch.

Change-Id: Ia1820495a989f4f84530ab83f2d87d53a9f761df