Report: Support Sequences and Mappings as Data
Previously, the ReportModel class assumed that data passed in
was a dict, without actually checking.  This could lead to errors
if anything else was passed in.  As it was likely one would want
to pass in arbitrary sequences (like lists) and mappings, support
was added.

It works as such:
- a MutableMapping is set directly as model.data
- a Mapping is wrapped with a COWMapping, which does copy-on-write
  to make a MutableMapping
- a MutableSequence is wrapped with a MutableSequenceToMappingProxy,
  which uses sequence indicies as keys
- a Sequence is wrapped with a SequenceToMappingProxy, which uses
  sequence indicies as key, and that is then wrapped with a
  COWMapping (as described above)

Finally, instances of manually converting an array to a dict were
removed.

Change-Id: Ibcf7d01d69ac21f9ad8c3dcf604733e88122c8f4