Report: Support Sequences and Mappings as Data
Previously, the ReportModel class assumed that data passed in
was a dict, without actually checking.  This could lead to errors
if anything else was passed in.  As it was likely one would want
to pass in arbitrary sequences (like lists) and mappings, support
was added.

It works as such:

- Mappings are transformed into dicts via the `dict` constructor

- Sequences are transformed into dicts by ziping the sequence
  together with its indicies, and then passing the result to
  the `dict` constructor.

Finally, instances of manually converting a list to a dict were
removed.

Change-Id: Ibcf7d01d69ac21f9ad8c3dcf604733e88122c8f4