Fix excessive logging from db.sqlalchemy.session
A message about the MySQL server mode was logged at the info level
every time a connection was checked out. This happens often, and
logging every time is unnecessary.

Here's an example of the message when using the default
configuration:

 MySQL server mode set to STRICT_TRANS_TABLES,STRICT_ALL_TABLES,...

The fix is to send the message once per engine creation.

Closes-Bug: #1288443
Co-Authored-By: Ben Nemec <bnemec@redhat.com>

Change-Id: I13b24d04c28e69a8534aaa8d3a4463ca1d65570f