Report: Fix JinjaView deepcopy
Previously, JinjaView used the default implementation of deepcopy.
Unfortunately, this meant that attempting to stringify the same model
with an attached JinjaView twice would fail, since Jinja's Template
class fails to deepcopy.

Now, a __deepcopy__ method has been explicitly writen for JinjaView
which reuses the Jinja Template (since two Templates for the same
text are logically going to be the same anyway).

Change-Id: Id040255138ecd48acd942d0d8ef75ed643e15403