Report: Propagate view_type through collections
Previously, if a model contained a property with a
collection (sequence or mapping) of models, set_current_view_type
would not propagate the current_view_type to those models.

This change allows the set_current_view_type to descend recursively
throught not just models, but also arbitrary sequences and mappings
to set the view type.

Change-Id: I3701b7b33195038ccc6b57456bfd0e732c32bbda