Change lazy translation to retain complete dict
When replacement text was retained for later translation, only
the entries that were associated with the keywords in the
format string were retained.  There was a bug in how the keywords
were parsed.  This reduction was done to support the use of
locals() to specify the replacement text.  This is now prevented
by H501 (confirmed in all projects).

This means that instead of fixing the parsing code to correctly
find the keywords, the entire dictionary can now be retained
and the code for finding the keywords and stripping out unused
entries removed.

Change-Id: I7969dc20841d839f68b2ddb29ac47340c6b615e8
Closes-Bug: #1288049