Don't store the request ID value in middleware
The original version of the request_id middleware generates a request ID during
process_request() and stores the value in a variable, so it can be attached as
header in process_request(). This is pretty dangerous, since subsequent
requests will overwrite this value, and the wrong request ID may be attached to
some responses. This is particularly apparent when requests arrive in parallel.

This change eliminates the process_response() block, leaving that work to the
applications using this middleware. With this middleware no longer tracking the
request ID, it is the responsibility of the application to pass the request ID
from the Request environ to the Response. While the application could pass the
value in with the Response for this middleware to handle, it's less convoluted
to just have the application attach the header itself.

Change-Id: I498859140ba113412c9d95d09f0c4c6b8650749c
Closes-bug: #1289696