Config generator: handle options in multiple groups
The commit 7c18261df5609d689a32c697c77c0a97512342bc disabled detection
of duplicate options in the same group, by comparing for equality rather
than identity. This had the effect of breaking the case where the same
option appeared in different groups, as it would always match the first
group the option appeared in, and thus show up only in that group.

This fixes the problem by only comparing for equality in the default
group. This is the place where it is most likely for duplicate options
to be registered in multiple different places. (It remains an open
question whether this should be allowed at all, or if bug 1262148 should
be fixed instead by eliminating duplicate options.) If duplicate options
are registered in any other group the original error-checking behaviour
will apply, since this ought never to happen.

Closes-bug: #1288586
Change-Id: Ie8753fbe84258e88b5a41fa7dccb45c224c69bde