Config generator: handle options in multiple groups
The commit 7c18261df5609d689a32c697c77c0a97512342bc disabled detection
of duplicate options in the same group, by comparing for equality rather
than identity. This had the effect of breaking the case where the same
option appeared in different groups, as it would always match the first
group the option appeared in, and thus show up only in that group. It
also had the affect that any options registered multiple times would
show up multiple times in one group in the output.

This fixes the problem by only matching on the identical object, while
still recognising if an equivalent option is registered and avoiding
raising the exception that caused bug 1262148 when that is the case.

Closes-bug: #1288586
Change-Id: Ie8753fbe84258e88b5a41fa7dccb45c224c69bde