Add an automated naming convention for UNIQUEs
For operating with UniqueConstraint names are needed, however these
names are in most cases generated by database in use, when the
constraint is created without name.

Omitting names when creating unique constraints leads to two problems:
1. UC name will be different for each database backend
(so we can't e.g. easily drop it in a migration script).
2. _raise_duplicate_error() depends on the specific format of UCs names
to return column names correctly.

Until we move to using SQLAlchemy 0.9, the implementation for
`remove_naming_conventions` can not be designed.

Change-Id: I5558ae6e7e1db2a50875641f551eef7360708e01