Memorycache: Make opt group configurable in get_client
The memorycache module registers the option "memcached_servers" in the
default group. However, the keystoneclient auth middleware also
registers the option in a separate group. To allow the auth middleware
to unregister the option from the default group, we need to allow the
caller of get_client() to pass the group in which to find the config -
CONF by default, but CONF.keystone_authtoken in the case of the auth
middleware.

Partial-bug: #1262148

Change-Id: I76e794a1dd8bcf580784f0b11fa184577ab0518b