Use 9001 instead of 9000 for test_pool
This patch uses the port 9001 instead of 9000 for the test_pool test. It
looks like something is running on the 9000 port on the gate and is
preventing this test to succeed. This test, as the others in this test
suite, are meant to test the server live, hence the listen method is not
being mocked out.

This test doesn't exist on master anymore, so this is a stable/havana
thing only.

Change-Id: I9d1f485244f852d7f3040d6464bb48e93382e8ed