Revert setting oslo-incubator logs to INFO
Originally oslo-incubator logs were set to INFO because we wanted to
treat oslo-incubator like a library, and we don't want debug logs on for
every single library that we use. In particular the lockutils module has
verbose logs. But this had the side effect of disabling the
processutils debug logs, which are very useful at letting you know when
a something is being shelled out to.

In the short term don't default oslo-incubator logs to INFO. And in the
long run we need to revisit how we set log levels, but that will have to
wait until Juno.

This is a partial revert of: Ia12c003dae196e2220a5f6ea768758a75f2ede23

Change-Id: Icc61964c6a2b002ac1c8566f174e6a5230674c75